src/requirements.txt
psutil>=5.6.2
pyax12>=0.5.dev3
websockets>=7.0
pyserial>=3.4
PySGP30
PyMLX90614
git+https://github.com/adafruit/Adafruit_AMG88xx_python#egg=Adafruit_AMG88xx
git+https://github.com/adafruit/Adafruit_bitfield_python#egg=Adafruit_bitfield
smbus2
multiprocessing-logging
seeed-python-mlx9064x
PyRoboteq
